Added better tcl syntax highlighting

+++ scripts/installers/tcl.sh
@@ -1,12 +1,30 @@
 #!/bin/bash
-TCL=tcl8.6.10
+
+# Builds Tcl/Tk from source and installs for current user
+
 TMP=$(mktemp -d)
 trap 'rm -rf $TMP' EXIT
-cd "$TMP" || exit 1
-wget "https://prdownloads.sourceforge.net/tcl/$TCL-src.tar.gz"
-tar -xf "$TCL-src.tar.gz"
-cd "$TCL/unix" || exit 1
-./configure --prefix="$HOME/.local"
-make
-make install
-ln -s "$HOME/.local/bin/tclsh8.6" "$HOME/.local/bin/tclsh"
+
+# Only add config flags to tk
+CONFIG=
+for TCL in tcl8.6.10 tk8.6.10; do
+    cd "$TMP" || exit 1
+    wget "https://prdownloads.sourceforge.net/tcl/$TCL-src.tar.gz"
+    tar -xf "$TCL-src.tar.gz"
+    cd "$TCL/unix" || exit 1
+    ./configure --prefix="$HOME/.local" $CONFIG
+    make
+    make install
+    CONFIG=--with-tcl="$TMP/tcl8.6.10/unix"
+done
+
+[ ! -e "$HOME/.local/bin/tclsh" ] && ln -s "$HOME/.local/bin/tclsh8.6" "$HOME/.local/bin/tclsh"
+
+# Install Nagelfar, a Tcl static analysis tool
+
+wget https://downloads.sourceforge.net/project/nagelfar/Rel_131/nagelfar131.tar.gz
+tar -xf nagelfar131.tar.gz
+cp -r nagelfar131 "$HOME/.local/"
+TARGET="$HOME/.local/bin/nagelfar"
+rm -f "$TARGET"
+ln -s "$HOME/.local/nagelfar131/nagelfar.tcl" "$TARGET"
